crdroid: Check if changelog file exists

+++ build/tools/changelog.sh
@@ -60,7 +60,10 @@ for i in $(seq $changelog_days); do
     echo >> $Changelog
 done
 
-sed -i 's/project/   */g' $Changelog
-
-cp $Changelog $OUT_DIR/target/product/$DEVICE/system/etc/
-mv $Changelog $OUT_DIR/target/product/$DEVICE/
+if [ -f $Changelog ]; then
+    sed -i 's/project/   */g' $Changelog
+    cp $Changelog $OUT_DIR/target/product/$DEVICE/system/etc/
+    mv $Changelog $OUT_DIR/target/product/$DEVICE/
+else
+    echo "Changelog file does not exist."
+fi
